requirements.txt
flask
psycopg2
flask_cors
python-decouple 
python-dotenv